// admin/src/generated/api/organization/models/warehouseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DCTD-UTC Commerce API
 * Contract for storefront and admin applications
 * OpenAPI spec version: 1.0.0
 */
import type { WarehouseDtoStatus } from './warehouseDtoStatus';

export interface WarehouseDto {
  id: string;
  branchId: string;
  code: string;
  name: string;
  status: WarehouseDtoStatus;
  isPrimary: boolean;
  version: number;
}